+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @repo/shared
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [114454a]
+- Updated dependencies [2cc7d49]
+- Updated dependencies [5964742]
+- Updated dependencies [0d71109]
+- Updated dependencies [06a4f14]
+  - @m2d/html@1.0.1
+
 ## 0.0.2
 
 ### Patch Changes